bổ sung

+++ ngo_van_luong/toan.h
@@ -2,4 +2,4 @@
 int cong(int a, int b);
 int tru(int a, int b);
 int nhan(int a, int b);
-int chia(int a, int b);
+float chia(float a,float b);
